+++ src/net_permissions.cpp
@@ -11,11 +11,13 @@
 
 const std::vector<std::string> NET_PERMISSIONS_DOC{
     "bloomfilter (allow requesting BIP37 filtered blocks and transactions)",
-    "noban (do not ban for misbehavior)",
+    "noban (do not ban for misbehavior; implies download)",
     "forcerelay (relay transactions that are already in the mempool; implies "
     "relay)",
     "relay (relay even in -blocksonly mode)",
     "mempool (allow requesting BIP35 mempool contents)",
+    "download (allow getheaders during IBD, no disconnect after "
+    "maxuploadtarget limit)",
 };
 
 namespace {
@@ -60,6 +62,8 @@ bool TryParsePermissionFlags(const std::string str, NetPermissionFlags &output,
                 NetPermissions::AddFlag(flags, PF_FORCERELAY);
             } else if (permission == "mempool") {
                 NetPermissions::AddFlag(flags, PF_MEMPOOL);
+            } else if (permission == "download") {
+                NetPermissions::AddFlag(flags, PF_DOWNLOAD);
             } else if (permission == "all") {
                 NetPermissions::AddFlag(flags, PF_ALL);
             } else if (permission == "relay") {
@@ -98,6 +102,9 @@ std::vector<std::string> NetPermissions::ToStrings(NetPermissionFlags flags) {
     }
     if (NetPermissions::HasFlag(flags, PF_MEMPOOL)) {
         strings.push_back("mempool");
+    }
+    if (NetPermissions::HasFlag(flags, PF_DOWNLOAD)) {
+        strings.push_back("download");
     }
     return strings;
 }